requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.11.3
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
httplib2==0.17.0
idna==2.9
logger==1.4
oauth2client==4.1.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2019.3
PyYAML==5.3
requests==2.23.0
rsa==4.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
